pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "funimage"
version = "1.0.14"
description = "A powerful Python library for image format conversion and processing"
readme = "README.md"
requires-python = ">=3.8"
keywords = [ "image", "conversion", "PIL", "opencv", "base64", "processing",]
classifiers = [ "Development Status :: 5 - Production/Stable", "Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Topic :: Software Development :: Libraries :: Python Modules", "Topic :: Multimedia :: Graphics",]
dependencies = [ "funfake>=1.0.3", "funutil>=1.0.14", "pillow>=9.0.0", "pillow-avif-plugin>=1.4.6", "numpy>=1.20.0", "requests>=2.25.0",]
[[project.authors]]
name = "牛哥"
email = "niuliangtao@qq.com"

[[project.authors]]
name = "farfarfun"
email = "farfarfun@qq.com"

[[project.maintainers]]
name = "牛哥"
email = "niuliangtao@qq.com"

[[project.maintainers]]
name = "farfarfun"
email = "farfarfun@qq.com"

[project.license]
text = "MIT"

[project.optional-dependencies]
opencv = [ "opencv-python>=4.5.0",]
dev = [ "pytest>=7.0.0", "pytest-cov>=4.0.0", "black>=22.0.0", "isort>=5.10.0", "flake8>=5.0.0", "mypy>=1.0.0",]

[project.urls]
Organization = "https://github.com/farfarfun"
Repository = "https://github.com/farfarfun/funimage"
Releases = "https://github.com/farfarfun/funimage/releases"

[tool.setuptools]
license-files = []


[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88


[tool.setuptools.package-dir]
"" = "src"


[tool.setuptools.packages.find]
where = [ "src",]
